# apps/control-api/src/app.ts
import {
  PresentationConflictError,
  PresentationUpdateError,
} from "@harbor-hf/control-core";
import { readFile } from "node:fs/promises";
import cookie from "@fastify/cookie";
import helmet from "@fastify/helmet";
import fastifyStatic from "@fastify/static";
import { ContractValidationError } from "@harbor-hf/contracts";
import { leaderboard } from "@harbor-hf/control-core";
import Fastify, {
  type FastifyInstance,
  type FastifyReply,
  type FastifyRequest,
} from "fastify";
import { z } from "zod";
import {
  type AuthenticatedActor,
  BearerRateLimitError,
  InvalidBearerCredentialError,
  OAuthCallbackError,
  type SessionRow,
} from "./auth.js";
import { HARBOR_REVISION } from "./harbor-revision.js";
import {
  HuggingFaceHardwareLookupError,
  lookupHuggingFaceHardware,
} from "./huggingface-hardware.js";
import {
  HuggingFaceModelLookupError,
  HuggingFaceModelNotFoundError,
  lookupHuggingFaceModelProviders,
} from "./huggingface-models.js";
import { LaunchError } from "./launch.js";
import type { Runtime } from "./runtime.js";
import { TrialProgressReader } from "./trial-progress-reader.js";

export { HARBOR_REVISION } from "./harbor-revision.js";

const providerSchema = z
  .string()
  .regex(
    /^[a-z0-9][a-z0-9-]{0,62}$/,
    "provider must use lowercase letters, numbers, and hyphens",
  );

const submissionSchema = z
  .object({
    benchmark: z
      .object({ name: z.string().min(1), preset: z.string().min(1) })
      .strict(),
    model: z
      .object({
        id: z.string().min(1).max(320),
        provider: providerSchema,
        reasoning_effort: z.string().min(1).max(40),
      })
      .strict(),
    harness: z
      .object({ agent: z.string().min(1), version: z.string().min(1) })
      .strict(),
    n_concurrent_trials: z.number().int().min(1).max(128).optional(),
    cost_ceiling_usd_per_trial: z.number().positive().max(10_000),
    role: z.enum(["final", "diagnostic"]).default("final"),
  })
  .strict();

const workbenchSubmissionSchema = submissionSchema
  .omit({ harness: true })
  .extend({
    model: z
      .object({
        id: z.string().min(1).max(320),
        provider: providerSchema,
        reasoning_effort: z.literal("off").default("off"),
      })
      .strict(),
    workbench: z
      .object({
        recipe: z.unknown(),
        setup_test_id: z.string().min(1).max(160),
        harbor_agent: z
          .object({ model_name: z.string().min(1).max(320) })
          .strict()
          .optional(),
      })
      .strict(),
  })
  .strict();

const workbenchSetupSchema = z.object({ recipe: z.unknown() }).strict();
const runParameters = z.object({ run_id: z.string().regex(/^run-[0-9a-f]{24}$/) });
const trialParameters = runParameters.extend({ trial_name: z.string().min(1) });
const setupParameters = z.object({ setup_test_id: z.string().min(1).max(160) });
const setupFileParameters = setupParameters.extend({
  file_id: z.string().min(1).max(160),
});
const modelProvidersQuery = z
  .object({ model: z.string().trim().min(1).max(320) })
  .strict();
const embeddedCookiePolicy = {
  httpOnly: true,
  secure: true,
  sameSite: "none" as const,
  partitioned: true,
};

interface RequestState {
  actor: AuthenticatedActor | null;
  session: SessionRow | null;
}

function state(request: FastifyRequest): RequestState {
  return (request as FastifyRequest & { controlState: RequestState }).controlState;
}

function error(reply: FastifyReply, status: number, code: string, message: string) {
  return reply.code(status).send({ error: { code, message } });
}

function contractValidationMessage(failure: ContractValidationError): string {
  const details = failure.errors
    .map(({ instancePath, keyword, message }) => {
      const path = instancePath || "request";
      return `${path}: ${message ?? keyword}`;
    })
    .join("\n");
  return details ? `${failure.message}:\n${details}` : failure.message;
}

async function authenticate(
  runtime: Runtime,
  request: FastifyRequest,
  reply: FastifyReply,
): Promise<boolean> {
  const requestState = state(request);
  if (runtime.config.auth_mode === "development") {
    requestState.actor = runtime.auth.developmentActor();
    return true;
  }
  const authorization = request.headers.authorization;
  if (authorization?.startsWith("Bearer ")) {
    try {
      requestState.actor = await runtime.auth.bearerActor(
        authorization.slice("Bearer ".length),
      );
      return true;
    } catch (failure) {
      if (failure instanceof BearerRateLimitError) {
        await reply.header("Retry-After", "60");
        error(reply, 429, "rate_limit_exceeded", "bearer lookup rate exceeded");
        return false;
      }
      if (!(failure instanceof InvalidBearerCredentialError)) throw failure;
      error(reply, 401, "unauthorized", "the bearer credential is invalid");
      return false;
    }
  }
  const sessionId = request.cookies.hhf_session;
  if (!sessionId) {
    error(reply, 401, "unauthorized", "sign in is required");
    return false;
  }
  const authenticated = await runtime.auth.sessionActor(sessionId);
  if (!authenticated) {
    error(reply, 401, "unauthorized", "the session is invalid or expired");
    return false;
  }
  requestState.actor = authenticated.actor;
  requestState.session = authenticated.session;
  if (!["GET", "HEAD"].includes(request.method)) {
    const token = request.headers["x-csrf-token"];
    if (
      typeof token !== "string" ||
      !runtime.auth.csrfValid(authenticated.session, token)
    ) {
      error(reply, 403, "csrf_rejected", "the CSRF token is invalid");
      return false;
    }
  }
  return true;
}

function requireActor(request: FastifyRequest): AuthenticatedActor {
  const actor = state(request).actor;
  if (!actor) throw new Error("authenticated actor is missing");
  return actor;
}

function idempotencyKey(request: FastifyRequest): string {
  const value = request.headers["idempotency-key"];
  if (typeof value !== "string" || !value.trim() || value.length > 320)
    throw new Error("Idempotency-Key header is required");
  return value;
}

function publicApi(path: string): boolean {
  return path === "/api/v1/leaderboard" || path === "/api/v1/session";
}

export async function buildApp(runtime: Runtime): Promise<FastifyInstance> {
  const progressReader = new TrialProgressReader(runtime.store);
  const app = Fastify({
    logger: runtime.config.node_env !== "test" && {
      // Allowlist request metadata: no query strings, headers, or cookies.
      serializers: {
        req: (request: FastifyRequest) => ({
          method: request.method,
          url: request.url.split("?", 1)[0] ?? "/",
        }),
      },
    },
  });
  await app.register(cookie);
  await app.register(helmet, {
    contentSecurityPolicy: {
      directives: {
        defaultSrc: ["'self'"],
        scriptSrc: ["'self'"],
        styleSrc: ["'self'"],
        imgSrc: ["'self'", "data:"],
        connectSrc: ["'self'"],
        objectSrc: ["'none'"],
        frameAncestors: ["'self'", "https://huggingface.co"],
      },
    },
    crossOriginEmbedderPolicy: false,
    crossOriginResourcePolicy: { policy: "cross-origin" },
    xFrameOptions: false,
  });
  app.decorateRequest("controlState", null);
  app.addHook("onRequest", async (request) => {
    (request as FastifyRequest & { controlState: RequestState }).controlState = {
      actor: null,
      session: null,
    };
  });
  app.addHook("preHandler", async (request, reply) => {
    const path = request.url.split("?", 1)[0] ?? request.url;
    if (!path.startsWith("/api/v1/") || publicApi(path)) return;
    if (!(await authenticate(runtime, request, reply))) return reply;
    const mutation = request.method !== "GET" && request.method !== "HEAD";
    if (mutation && requireActor(request).role !== "operator")
      return error(reply, 403, "operator_required", "operator access is required");
    const workbenchPreview =
      path === "/api/v1/workbench/preview" || path === "/api/v1/runs/validate";
    const localWorkbench =
      path.startsWith("/api/v1/workbench/setup-tests") &&
      runtime.config.workbench_runner === "docker";
    const workbenchCancel =
      path.startsWith("/api/v1/workbench/setup-tests/") && path.endsWith("/cancel");
    if (
      mutation &&
      runtime.config.write_mode !== "enabled" &&
      !workbenchPreview &&
      !localWorkbench &&
      !workbenchCancel
    )
      return error(reply, 503, "write_disabled", "write mode is disabled");
  });

  app.setErrorHandler((failure, request, reply) => {
    if (reply.sent) return;
    if (failure instanceof PresentationConflictError)
      return error(reply, 409, "presentation_conflict", failure.message);
    if (failure instanceof PresentationUpdateError) {
      request.log.error(
        { code: "presentation_update_failed" },
        "Archive update failed",
      );
      return error(reply, 503, "presentation_update_failed", failure.message);
    }
    if (failure instanceof OAuthCallbackError) {
      const code = failure.denied ? "access_denied" : "oauth_failed";
      request.log.error({ oauth_stage: failure.stage, code }, "OAuth callback failed");
      const status = failure.denied
        ? 403
        : failure.stage === "flow" || failure.stage === "token_exchange"
          ? 400
          : 500;
      return error(reply, status, code, failure.message);
    }
    if (failure instanceof HuggingFaceHardwareLookupError)
      return error(reply, 503, "hardware_unavailable", failure.message);
    if (failure instanceof LaunchError)
      return error(
        reply,
        failure.status,
        failure.status === 400 ? "invalid_request" : "inspection_unavailable",
        failure.message,
      );
    if (failure instanceof z.ZodError)
      return error(
        reply,
        400,
        "invalid_request",
        `the request is invalid:\n${z.prettifyError(failure)}`,
      );
    if (failure instanceof ContractValidationError)
      return error(reply, 400, "invalid_request", contractValidationMessage(failure));
    if (failure instanceof HuggingFaceModelNotFoundError)
      return error(reply, 404, "model_not_found", failure.message);
    if (failure instanceof HuggingFaceModelLookupError)
      return error(reply, 502, "hub_lookup_failed", failure.message);
    const message = failure instanceof Error ? failure.message : "request failed";
    if (message.includes("not found")) return error(reply, 404, "not_found", message);
    if (message.includes("already identifies") || message.includes("cancelled run"))
      return error(reply, 409, "conflict", message);
    if (
      /Idempotency-Key|preset|reasoning|cost ceiling|JobConfig|credential literal|environment|agent|Workbench|workbench|setup test|recipe/.test(
        message,
      )
    )
      return error(reply, 400, "invalid_request", message);
    app.log.error(
      { error_name: failure instanceof Error ? failure.name : "Error" },
      "request failed",
    );
    return error(reply, 500, "internal_error", "the request failed");
  });

  app.get("/health/live", async () => ({ status: "live" }));
  app.get("/health/ready", async (_request, reply) =>
    runtime.ready
      ? { status: "ready" }
      : reply.code(503).send({ status: "initializing" }),
  );

  app.get("/auth/login", async (request, reply) => {
    if (runtime.config.auth_mode === "development") return reply.redirect("/");
    const query = z
      .object({ return_to: z.string().optional().default("/") })
      .parse(request.query);
    const login = await runtime.auth.login(query.return_to);
    reply.setCookie("hhf_oauth_flow", login.flow_id, {
      ...embeddedCookiePolicy,
      path: "/auth/callback",
      maxAge: 600,
    });
    return reply.redirect(login.url.toString());
  });

  app.get("/auth/callback", async (request, reply) => {
    const flowId = request.cookies.hhf_oauth_flow;
    if (!flowId) throw new OAuthCallbackError("flow");
    const callback = await runtime.auth.callback(
      flowId,
      new URL(request.url, runtime.config.public_origin),
    );
    reply.clearCookie("hhf_oauth_flow", {
      ...embeddedCookiePolicy,
      path: "/auth/callback",
    });
    reply.setCookie("hhf_session", callback.session_id, {
      ...embeddedCookiePolicy,
      path: "/",
      expires: new Date(callback.expires_at),
    });
    reply.setCookie("hhf_csrf", callback.csrf, {
      ...embeddedCookiePolicy,
      httpOnly: false,
      path: "/",
      expires: new Date(callback.expires_at),
    });
    return reply.redirect(callback.return_to);
  });

  app.post("/auth/logout", async (request, reply) => {
    const sessionId = request.cookies.hhf_session;
    if (sessionId) runtime.auth.store.deleteSession(sessionId);
    reply.clearCookie("hhf_session", { ...embeddedCookiePolicy, path: "/" });
    reply.clearCookie("hhf_csrf", { ...embeddedCookiePolicy, path: "/" });
    return reply.code(204).send();
  });

  app.get("/api/v1/session", async (request, reply) => {
    if (!(await authenticate(runtime, request, reply))) return reply;
    const actor = requireActor(request);
    return {
      authenticated: true,
      actor: { username: actor.username, role: actor.role, transport: actor.transport },
    };
  });

  app.get("/api/v1/system", async () => ({
    source_revision: runtime.config.source_revision,
    harbor_revision: HARBOR_REVISION,
    write_mode: runtime.config.write_mode,
    ready: runtime.ready,
    projection: runtime.projection.system(),
    capacity: { max_active_parent_jobs: runtime.config.max_active_jobs },
    workbench: {
      runner: runtime.config.workbench_runner,
      setup_enabled: runtime.config.workbench_runner !== "disabled",
    },
    resources: { spaces: 1, buckets: 1, operator_secrets: 2 },
  }));

  app.get("/api/v1/presets", async () => ({
    benchmarks: runtime.presets.benchmarks,
    agents: runtime.presets.agents,
  }));

  app.get("/api/v1/agents", async () => runtime.launch.catalog());

  app.post("/api/v1/runs/validate", async (request) =>
    runtime.launch.validate(request.body),
  );

  app.get("/api/v1/model-providers", async (request) => {
    const { model } = modelProvidersQuery.parse(request.query);
    return {
      model,
      providers: await lookupHuggingFaceModelProviders(model),
    };
  });

  app.post("/api/v1/workbench/preview", async (request) =>
    runtime.workbench.preview(request.body),
  );

  app.post("/api/v1/workbench/setup-tests", async (request, reply) => {
    const actor = requireActor(request);
    const input = workbenchSetupSchema.parse(request.body);
    const setup = await runtime.workbench.startSetup(
      input.recipe,
      actor.subject,
      idempotencyKey(request),
    );
    return reply.code(202).send(setup);
  });

  app.get("/api/v1/workbench/setup-tests", async (request) => ({
    setups: await runtime.workbench.listSetups(requireActor(request).subject),
  }));

  app.get("/api/v1/workbench/setup-tests/:setup_test_id", async (request) => {
    const { setup_test_id } = setupParameters.parse(request.params);
    const setup = await runtime.workbench.getSetup(
      setup_test_id,
      requireActor(request).subject,
    );
    if (!setup) throw new Error("setup test was not found");
    return setup;
  });

  app.post("/api/v1/workbench/setup-tests/:setup_test_id/cancel", async (request) => {
    const { setup_test_id } = setupParameters.parse(request.params);
    const setup = await runtime.workbench.cancelSetup(
      setup_test_id,
      requireActor(request).subject,
    );
    if (!setup) throw new Error("setup test was not found");
    return setup;
  });

  app.get("/api/v1/workbench/setup-tests/:setup_test_id/logs", async (request) => {
    const { setup_test_id } = setupParameters.parse(request.params);
    const logs = await runtime.workbench.logs(
      setup_test_id,
      requireActor(request).subject,
    );
    if (!logs) throw new Error("setup test was not found");
    return logs;
  });

  app.get(
    "/api/v1/workbench/setup-tests/:setup_test_id/files/:file_id",
    async (request) => {
      const { setup_test_id, file_id } = setupFileParameters.parse(request.params);
      const file = await runtime.workbench.file(
        setup_test_id,
        file_id,
        requireActor(request).subject,
      );
      if (!file) throw new Error("Workbench file was not found");
      return file;
    },
  );

  app.post("/api/v1/runs", async (request, reply) => {
    const actor = requireActor(request);
    const body = request.body as Record<string, unknown>;
    if (body && typeof body === "object" && "workbench" in body) {
      const input = workbenchSubmissionSchema.parse(body);
      const preview = runtime.workbench.preview(input.workbench.recipe);
      const sources = new Set(preview.recipe.environment.map((item) => item.source));
      if (!sources.has("model_base_url") || !sources.has("model_api_key"))
        throw new Error(
          "Workbench Run requires model_base_url and model_api_key bindings",
        );
      await runtime.workbench.attestPassedSetup(
        input.workbench.setup_test_id,
        actor.subject,
        input.workbench.recipe,
      );
      const result = await runtime.service.submitWorkbench(
        {
          benchmark: input.benchmark,
          model: input.model,
          harness: { agent: "command-agent", version: preview.revision_id },
          n_concurrent_trials: input.n_concurrent_trials,
          cost_ceiling_usd_per_trial: input.cost_ceiling_usd_per_trial,
          role: input.role,
        },
        { ...preview.harbor_agent, ...input.workbench.harbor_agent },
        idempotencyKey(request),
        actor.subject,
        { name: preview.recipe.name },
      );
      return reply.code(result.created ? 201 : 200).send(result);
    }
    const result = await runtime.service.submitPreset(
      submissionSchema.parse(body),
      idempotencyKey(request),
      actor.subject,
    );
    return reply.code(result.created ? 201 : 200).send(result);
  });

  app.post("/api/v1/runs/config", async (request, reply) => {
    const actor = requireActor(request);
    const key = idempotencyKey(request);
    const ceiling = Number(request.headers["x-harbor-hf-cost-ceiling-usd-per-trial"]);
    if (!Number.isFinite(ceiling) || ceiling <= 0 || ceiling > 10_000)
      throw new Error("cost ceiling must be a finite positive USD value");
    const validation = await runtime.launch.validate(request.body);
    if (!validation.credentials_available)
      return error(
        reply,
        503,
        "credentials_unavailable",
        "Separate control and inference credentials are required to launch",
      );
    const reviewed = request.headers["x-harbor-hf-validation"];
    if (reviewed !== undefined && reviewed !== validation.fingerprint)
      return error(
        reply,
        409,
        "validation_changed",
        "Launch configuration or admission policy changed; validate again",
      );
    const result = await runtime.service.submitConfig(
      request.body,
      ceiling,
      key,
      actor.subject,
    );
    return reply.code(result.created ? 201 : 200).send(result);
  });

  app.get("/api/v1/runs", async () => ({ runs: runtime.projection.listRuns() }));
  app.get("/api/v1/runs/:run_id", async (request) => {
    const { run_id } = runParameters.parse(request.params);
    const run = runtime.projection.run(run_id);
    if (!run) throw new Error("run was not found");
    return run;
  });

  app.patch("/api/v1/runs/:run_id/presentation", async (request) => {
    const actor = requireActor(request);
    const { run_id } = runParameters.parse(request.params);
    const input = z
      .strictObject({
        archived: z.boolean(),
        expected_revision: z.number().int().min(0).max(Number.MAX_SAFE_INTEGER),
      })
      .parse(request.body);
    return runtime.service.setPresentation(
      run_id,
      input.archived,
      input.expected_revision,
      actor.subject,
    );
  });

  for (const [action, desired] of [
    ["pause", "paused"],
    ["resume", "run"],
    ["cancel", "cancelled"],
  ] as const) {
    app.post(`/api/v1/runs/:run_id/${action}`, async (request) => {
      const actor = requireActor(request);
      const { run_id } = runParameters.parse(request.params);
      return runtime.service.setDesiredState(run_id, desired, actor.subject);
    });
  }

  app.get("/api/v1/hardware", async () => lookupHuggingFaceHardware());

  app.get("/api/v1/runs/:run_id/progress", async (request) => {
    const { run_id } = runParameters.parse(request.params);
    if (!runtime.projection.run(run_id)) throw new Error("run was not found");
    const observation = runtime.projection.jobObservations();
    return progressReader.snapshot(run_id, observation.jobs, observation.observed_at);
  });

  app.get("/api/v1/runs/:run_id/trials", async (request) => {
    const { run_id } = runParameters.parse(request.params);
    if (!runtime.projection.run(run_id)) throw new Error("run was not found");
    return { trials: runtime.projection.trials(run_id, "identity") };
  });

  app.get("/api/v1/runs/:run_id/trials/:trial_name", async (request) => {
    const { run_id, trial_name } = trialParameters.parse(request.params);
    const trial = runtime.projection
      .trials(run_id)
      .find((item) => item.trial_name === trial_name);
    if (!trial) throw new Error("trial was not found");
    return trial;
  });

  app.get("/api/v1/jobs", async () => ({ jobs: runtime.projection.jobs() }));
  app.get("/api/v1/leaderboard", async () => ({
    rows: leaderboard(runtime.projection, runtime.presets),
  }));

  await app.register(fastifyStatic, {
    root: runtime.config.web_root,
    wildcard: false,
  });
  app.setNotFoundHandler(async (request, reply) => {
    if (request.url.startsWith("/api/") || request.url.startsWith("/auth/"))
      return error(reply, 404, "not_found", "route was not found");
    if (request.method !== "GET")
      return error(reply, 404, "not_found", "route was not found");
    const index = await readFile(`${runtime.config.web_root}/index.html`, "utf8");
    return reply.type("text/html; charset=utf-8").send(index);
  });
  return app;
}